+++ src/main.tsx
@@ -1,6 +1,6 @@
 import CircularProgress from '@mui/material/CircularProgress'
 import { HoxRoot } from 'hox'
-import { Suspense } from 'react'
+import { StrictMode, Suspense } from 'react'
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 import 'virtual:windi.css'
@@ -9,19 +9,21 @@ import { GlobalStyles } from './globalStyles'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-  <HoxRoot>
-    <BrowserRouter>
-      <Suspense
-        fallback={
-          <div className="w-screen h-screen fjic">
-            <CircularProgress />
-          </div>
-        }
-      >
-        <GlobalStyles>
-          <App />
-        </GlobalStyles>
-      </Suspense>
-    </BrowserRouter>
-  </HoxRoot>
+  <StrictMode>
+    <HoxRoot>
+      <BrowserRouter>
+        <Suspense
+          fallback={
+            <div className="w-screen h-screen fjic">
+              <CircularProgress />
+            </div>
+          }
+        >
+          <GlobalStyles>
+            <App />
+          </GlobalStyles>
+        </Suspense>
+      </BrowserRouter>
+    </HoxRoot>
+  </StrictMode>
 )
